wallet: move tests into test/ dir like other unit tests.
Trivial move, but makes it easy to add new ones.
