+++ wallet/test/.gitignore
@@ -1,0 +1,2 @@
+run-db
+run-wallet
