+++ wallet/test/Makefile
@@ -1,0 +1,27 @@
+WALLET_TEST_SRC := $(wildcard wallet/test/run-*.c)
+WALLET_TEST_OBJS := $(WALLET_TEST_SRC:.c=.o)
+WALLET_TEST_PROGRAMS := $(WALLET_TEST_OBJS:.o=)
+
+WALLET_TEST_COMMON_OBJS :=				\
+	common/htlc_state.o			\
+	common/type_to_string.o			\
+	common/pseudorand.o			\
+	common/utils.o				\
+	lightningd/htlc_end.o			\
+	lightningd/log.o
+
+# Make sure these depend on everything.
+ALL_TEST_PROGRAMS += $(WALLET_TEST_PROGRAMS)
+ALL_OBJS += $(WALLET_LIB_OBJS) $(WALLET_TEST_OBJS)
+
+$(WALLET_TEST_PROGRAMS): $(BITCOIN_OBJS) $(WALLET_TEST_COMMON_OBJS)
+$(WALLET_TEST_OBJS): $(WALLET_LIB_HEADERS)
+
+wallet/tests: $(WALLET_TEST_PROGRAMS:%=unittest/%)
+
+$(WALLET_TEST_PROGRAMS): $(WALLET_TEST_COMMON_OBJS) $(BITCOIN_OBJS)
+$(WALLET_TEST_OBJS): $(WALLET_SRC)
+
+update-mocks: $(WALLET_TEST_SRC:%=update-mocks/%)
+
+check: $(WALLET_TEST_PROGRAMS:%=unittest/%)
